Bug fixes and refactoring
Bug fixes
---------
Lines 300 and 335 in player25.py were crashing with "TypeError: a
bytes-like object is required, not 'str'".  The output of
subprocess.check_output() needed to be converted from bytes to string
before string operations could happen.

Refactoring
-----------
Refactored Piradio.desktop to not use hardcoded paths to a home
directory that may or may not exist.  This schortcut will reference the
home directory of the user who launched it.